+++ SignalServiceKit/Groups/GroupsV2AvatarDownloadOperation.swift
@@ -6,7 +6,7 @@
 import Foundation
 
 enum GroupsV2AvatarDownloadOperation {
-    static func run(urlPath: String, maxDownloadSize: UInt) async throws -> Data {
+    static func run(urlPath: String, maxDownloadSize: UInt64) async throws -> Data {
         return try await Retry.performWithBackoff(maxAttempts: 4) {
             return try await CDNDownloadOperation.tryToDownload(urlPath: urlPath, maxDownloadSize: maxDownloadSize)
         }
